Update 03_report-tool-architecture-beyond-models.tex
Handled the TODO on the study types. IMHO, it is actually quite generic for all study types, as long as they involve other components. Tried to emphasize this, connecting it to the new introduction on the "system-level". I also changed "developing" to "involving" so that it also covers researchers using LLM-based systems in their studies. Another minor change is that I generalized RAG to retrieval-augmented methods.